+++ build_runner/lib/src/build_runner.dart
@@ -11,7 +11,7 @@ import 'package:io/io.dart';
 import 'package:path/path.dart' as p;
 import 'package:yaml/yaml.dart';
 
-import 'bootstrap/bootstrap.dart';
+import 'bootstrap/bootstrapper.dart';
 import 'build_plan/build_options.dart';
 import 'build_plan/builder_factories.dart';
 import 'build_runner_command_line.dart';
@@ -160,15 +160,9 @@ class BuildRunner {
       b.mode = commandLine.type.buildLogMode;
     });
 
-    // Build and run, retrying if the nested `build_runner` invocation exits
-    // with exit code `tempFail`.
-    while (true) {
-      exitCode = await generateAndRun(
-        arguments,
-        experiments: commandLine.enableExperiments,
-      );
-
-      if (exitCode != ExitCode.tempFail.code) return exitCode;
-    }
+    return await Bootstrapper().run(
+      arguments,
+      experiments: commandLine.enableExperiments,
+    );
   }
 }