+++ build_runner/test/commands/serve/asset_handler_test.dart
@@ -27,7 +27,6 @@ void main() {
     assetGraph = await AssetGraph.build(
       BuildPhases([]),
       <AssetId>{},
-      <AssetId>{},
       buildPackageGraph({rootPackage('a'): []}),
       InternalTestReaderWriter(),
     );