+++ build_runner/CHANGELOG.md
@@ -1,12 +1,12 @@
 ## 2.8.1-wip
 
-- Watch mode: handle `build.yaml` changes without restarting.
+- Rewrite bootstrap code to remove use of `dart:mirrors`.
+- Watch mode: handle builder code and config changes without exiting.
 - Remove log output about `build_runner` internals.
 - Print the port that gets picked if you pass 0 for a port number, for example
   with `dart run build_runner serve web:0`.
 - Improved warnings when an option is specified for an unknown builder.
 - Bug fix: require `args` 2.5.0.
-- Use `build` ^4.0.0.
 
 ## 2.8.0